requirements.txt
numpy==1.23.5
pillow==9.5.0
scipy==1.13.0
scikit-image==0.21.0
scikit-learn==1.3.1
pycocotools==2.0.7
nltk==3.8.1
torch==2.0.1
torchvision==0.15.2
mmagic
salesforce-lavis
einops==0.4.1
pytorch-lightning==1.9.4
accelerate==0.21.0
diffusers==0.22.3
mmcv==2.0.0
gradio==4.37.1
spacy
opencv-python==4.9.0.80
transformers==4.30.2
basicsr
clip
open_clip_torch
